requirements.txt
matplotlib>=3.2.1
numpy>=1.18.2
Pillow>=8.3.2
cairosvg>=2.5.2
